Create NetworkData UseCases

+++ src/com/example/network/statistic/db/Db.kt
@@ -16,7 +16,7 @@ object Db {
     object NetworkData : Table() {
         val id = integer("id").autoIncrement().primaryKey() // Column<String>
         val userId = varchar("user_id", 50).index()
-        val timestamp = integer("timestamp").index()
+        val timestamp = long("timestamp").index()
         val period =varchar("period", 10)
         val data = varchar("data", 300)
     }
